requirements.txt
aiogram>=3.0.0
yt-dlp>=2023.10.13
python-dotenv>=1.0.0
